WIP-2 PRODX-19642 Updated after review. Move Cloud and store it properly in CloudStore
